Well, this will do for now. Obviously I won't be starting out today swinging. It will take me a bit to get into Python.
I'll have to learn how the RL environments and PyTorch work as well which will take me a bit.

For my first goal, I think I should train a feedforward net on Mnist just to get a feel for it.

Once I am at that point it won't be too hard to move to bigger and better things.

The reason why the whole front end issue that was plaguing me in Spiral is so complicated is because if the barely started thing I did then is so hard, then just how much difficulty am I going to have once I start making GUIs and parallelizing the environments? How long would the compile times be for that?

I do not want to find out.

The continuing development that other teams are doing is putting pressure on me as well.

True mastery will require a more open approach than the one I've been doing now, so let me start it here with this new repo.

As I go forward I'll look for ways extending dynamic languages with the benefits of intensional typing.