"10:55am.
```
    def forward(self, x):
        x = self.pool(F.relu(self.conv1(x)))
        x = self.pool(F.relu(self.conv2(x)))
        x = x.view(-1, 16 * 5 * 5)
        x = F.relu(self.fc1(x))
        x = F.relu(self.fc2(x))
        x = self.fc3(x)
        return x
```

This pattern. I've seen it everywhere and it is pretty much a symptom of bad programming. They really should just put it in a list and fold over it. Or alternatively use the pipe operator which Python does not have. But this would be easy to work around.

It is really amazing how easy things seem to be in Python. Like printing out the images to screen.

11:25am. I've gone through the first PyTorch tutorial (again). Simply put, the way Python does it, by having an interactive shell and allowing me to run each part step by step is the way this should be done. This takes full advantage of the strengths of dynamic typing and I could only dream of having this functionality in Spiral.

One thing I've noticed is that tensors seem to have the wrong order. The inner dimension in Python seems to come first.

This is a design mistake, but it is not a deal breaker.

11:30am. This is a bit fun. Yeah, there is a lot I could do with this. At the end of this journey I will put all the pieces together eventually, but now I am cursed to wander the world.

https://pytorch.org/tutorials/beginner/data_loading_tutorial.html

I'll go through this next. The ability to just display images in some window would be to kill for in Spiral.

I actually like that Python uses reference counting and will immediately free memory when a cell goes out of scope. I'd much rather prefer that to having to mess around with regions."