"2:05pm.
```
model = torch.nn.Sequential(
    torch.nn.Linear(D_in, H),
    torch.nn.ReLU(),
    torch.nn.Linear(H, D_out),
)
```

I wonder if there is way to avoid having to put in intermediate sizes like in Spiral? If there is not, it should not be hard to implement.

2:20pm. https://pytorch.org/tutorials/beginner/examples_autograd/tf_two_layer_net.html#sphx-glr-beginner-examples-autograd-tf-two-layer-net-py

I admit that it took me a bit of staring to figure out just where the assignment is happening in the TF example. I really favor the autograd example quite a bit more than TFs approach.

2:35pm. https://www.programiz.com/python-programming/decorator

Oh, lol. I did not expect decorators to be this.

2:50pm. https://pytorch.org/tutorials/beginner/examples_nn/two_layer_net_module.html

So PyTorch modules really are not using partial application to separate the input and hidden sizes. That is a design miss.

I wonder how the optimizer inspects the object in order to extract the parameters.

I wonder if it is possible to define a class inside a function.

```
def test():
    class Qwe:
        pass
    return Qwe

a = test()
```

Oh, yes it is. I remember wanting F# of being able to do this.

3:05pm. https://pytorch.org/tutorials/beginner/transfer_learning_tutorial.html

I'll skip this one.

https://pytorch.org/tutorials/beginner/deploy_seq2seq_hybrid_frontend_tutorial.html

And this one, but let me just see what the hybrid frontend actually is.

```
Although the eager interface is a beneficial tool for research and experimentation applications, when it comes time to deploy the model in a production environment, having a graph-based model representation is very beneficial. A deferred graph representation allows for optimizations such as out-of-order execution, and the ability to target highly optimized hardware architectures. Also, a graph-based representation enables framework-agnostic model exportation. PyTorch provides mechanisms for incrementally converting eager-mode code into Torch Script, a statically analyzable and optimizable subset of Python that Torch uses to represent deep learning programs independently from the Python runtime.
```

https://pytorch.org/docs/master/jit.html

Ah, this is definitely of interest to me.

3:20pm. Change of plans, I'll go through the entire reference before moving forward. In terms of capability Torch Script seems sharply inferior to Spiral, but the foundation they are setting up is what might allow me to make Spiral for PyTorch.

3:55pm. Done with chores. Let me resume. I'll get through the reference and take a break...actually, let me have lunch now instead. I usually have it at 6pm, but better to have it now than when it cools.

4:45pm. Done with the JIT reference. I thought it would be larger, 4/5ths of it is just the function list.

https://pytorch.org/tutorials/beginner/saving_loading_models.html

Let me go through this and the next chapter and then I will call it a day. I do not feel like doing too much today. I am near the end of my focus anyway.

5:10pm. The last chapter was just a recap of what came before so I skimmed it.

https://pytorch.org/tutorials/intermediate/reinforcement_q_learning.html

Interesting that the use Huber loss here. I want to say that this is a good idea, but with proper reward scaling it won't be necessary.

That having said this is one way I could cover huge reward scales that would otherwise destabilize the squared error.

5:20pm. I'll stop here for the day.

Tomorrow I will play with the RL example a little and then take a look at the OpenAI gym. After that I'll give implementing TD+ a try.

As for the JIT, let me put in a few words...I can definitely beat this though I am not sure how far I'd need to go. Should I reimplement Spiral in Python? I'd rather not mess with that sort of thing. I'd probably need to change the language so it produces C++ on host as well.

It is not a priority either way. I'll just end up using KFAC + vanilla NNs for the time being anyway."