https://www.youtube.com/playlist?list=PLbMVogVj5nJQNjkHZgwuAlfQ9tzmQDxjA Computer Science - Principles of Compiler Design
https://www.youtube.com/playlist?list=PLEbnTDJUr_IcPtUXFy2b1sGRPsLFMghhS
Compiler Design

https://www.youtube.com/playlist?list=PL0s3O6GgLL5fLboFu3TwM8XCgbm7m5U5n
System Programming and Compiler Construction Lectures

Right now, I should be playing with the RL example, but I miss Spiral. I've looked through some of Python's source code, but it is just a big clump of branches - no way I could figure that out. That sort of thing can only evolve, but in order to learn I need to internalize the design.

One big failure of Spiral that falls on me is that I never got a proper grip not on just its compile time performance, but even how to properly measure it.

An episode come to mind back in 2017 when I was working on its parser library. For some reason the static printf really took a long time to compile. Like 0.1s for a single print statement. I ended up just blaming it on monads and forgetting about it, but now I've gone to gripping about how a single gemm call increases compilation time by a similar amount.

...

What is the right choice here?

I could print out the trace and see exactly what path is visiting...but after that what then. The issue is, I sort of have a sense of where overheads lie, but how would I speed up that parser for example?

...

I have to dwell on this a bit. Just what makes for an efficient compiler?