"2:05pm. I finally said my goodbye to Spiral and the rain - it is falling. I never want to leave anything ever again.
If the Lispers and other metaprogrammers were more capable this would never have had to happen. But if that were the case, I would never have needed to create Spiral in the first place. My vision is strong enough to unify all the aspects of programming, but I do not have the power to see it through to completion. If I get better at machine learning, I might just one day get to the point where it will be possible.

Well Python, for the time being I will be in your care. Sorry for looking down on you all this time. I was wrong - mastering interpreted dynamic languages is definitely an important aspect of mastering programming as a whole.

Even if one prioritizes runtime performance, the compilation performance for such a language would require dynamism under the hood. It is impossible to avert from.

2:15pm. I've almost been programming for 4 years at this point. This is about the time the average person goes to high school.

Surely I learned a few things that will help me when programming in a dynamic language such as Python.

I've been wondering all this time. Am I an idiot for trying to strike out on my own, or have a gained something that will make a decisive difference?

2:25pm. I must be in hell. As I go into RL I will be using the wrong algorithm, the wrong language and the things I love will not be the things that give me power. I am completely the wrong person to cause the Singularity, but on the other hand I do not believe anyone better exists in this world.

Maybe that is what makes the story interesting to those watching from beyond?

I cannot complain as the setup is what ensures my existence.

When the end of the era comes, will it be despair or hope?

The only thing going in my favor is that in the last 4 years I gained just a little bit of skill. And the field moved a little bit as well. So it will have to be done - I must test out the TD+ idea and make my way from there toward justice. I am looking forward to seeing what is the power of dynamic languages now that I've stretched Spiral past its breaking point.

2:30pm. I am stressed out at the moment. Rather than starting, I think I will step away from the screen for a bit.

I've spent the whole night thinking about Spiral, I did not think I would just drop all that work and am not mentally prepared to start doing ML."