I know this is resting time, but I just want to rant a bit.
Back in 2018 I thought that Spiral could surpass all given that is has higher expressiveness than extensionaly typed languages and higher performance than dynamically typed ones, but I was not counting on the problem with GCing GPU memory, the slowness of the compiler or the sheer difficulty of programming in it.

I thought abstraction would be the cure for all of that, but it is not enough.

I spent two years walking through a desert seemingly without end and now that I am on the verge of reaching my goal, in front of me is a fortress full monsters, barring the way forward.

It is just not enough. My abilities are simply not enough for this challenge. At best I did like half of the Spiral language.

I made a mistake and will need the power of the future to bail me out. I have no choice but to believe in my skills and tap into the power of dynamic languages if I want to survive.

It would be different if I started work on this in 2007, but right now it is just too late unless I get a lease on life.

The time to act is now, at any cost. I can no longer afford the productivity hits that come from doing it directly in Spiral.

Even though I love it and see it as missing piece in PL design, the time has come to either ride the wave or be swept away.

In order to attain perfection all the pieces will eventually have to come together. Spiral will get its chance. I'll have to believe in that and roll the dice. Whatever they land on, the result will be much better than not rolling at all.

...

Tomorrow I will steel my mind and go through the PyTorch RL example. After that comes the OpenAI gym or some other environment. After that I'll start work on TD+. That is the plan.

Though I am behind schedule, I think it should not be too hard to build things up quickly and start making inroads into RL. Especially given the sheer expressiveness of dynamic languages. After 4 years of practice, I think I should now have the skill to take advantage of them in earnest. My complaints about Python being too imperative or not having pattern matching are irrelevant. It does not matter that it is tilted towards imperative programming given that it has first class functions. And the lack of pattern matching can be mitigated by using a OO message passing style. Those weird scoping rules can be mitigated by simply using references for assignment.

The language is weird mishmash of features like list comprehensions, generators and various other things which should not be built in, but implemented directly, but it does not matter. It is not like those things get in the way.

There are some things that I like about Python besides its interactivity, reference counting and size of ecosystem. Though it does not have pattern matching, it does have default arguments in functions and its syntax is not too bad.

Default arguments are kind of a grey area in Spiral, they can be done using pattern matching, but in various places it should be possible to make a mistake by misnaming when using modules. This is not the case with Python.

Spiral is also too opaque, as a side effect of it expressiveness. Python on the other hand is quite verbose, but plays it to its advantage.

Well, that is all I wanted to get out my system for now.