"9:20am.
```
# if gpu is to be used
device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
```

So Python does have those expression based if statements. Unfortunately the syntax for them is confusing.

```
Transition = namedtuple('Transition',
                        ('state', 'action', 'next_state', 'reward'))
```

This is quite interesting functionality. I really want to know what the `**` operator does in Python.

https://stackoverflow.com/questions/36901/what-does-double-star-asterisk-and-star-asterisk-do-for-parameters

9:25am. Ok, I was wrong about Python not having pattern matching. It has it, just not in a form I ever expected.

```
first, *l, last = [1,2,3,4]
first, q, w, last = [1,2,3,4]
```

Both of these work as one would expect.

9:45am.

```

def select_action(state):
    global steps_done
    sample = random.random()
    eps_threshold = EPS_END + (EPS_START - EPS_END) * math.exp(-1. * steps_done / EPS_DECAY)
    steps_done += 1
    if sample > eps_threshold:
        with torch.no_grad():
            # t.max(1) will return largest value for column of each row.
            # second column on max result is index of where max element was
            # found, so we pick action with the larger expected reward.
            return policy_net(state).max(1)[1].view(1, 1)
    else:
        return torch.tensor([[random.randrange(2)]], device=device, dtype=torch.long)
```

I managed to fix this function.

```
def select_action(state):
    global steps_done
    sample = random.random()
    eps_threshold = EPS_END + (EPS_START - EPS_END) * \
        math.exp(-1. * steps_done / EPS_DECAY)
    steps_done += 1
    if sample > eps_threshold:
        with torch.no_grad():
            # t.max(1) will return largest value for column of each row.
            # second column on max result is index of where max element was
            # found, so we pick action with the larger expected reward.
            return policy_net(state).max(1)[1].view(1, 1)
    else:
        return torch.tensor([[random.randrange(2)]], device=device, dtype=torch.long)
```

For some reason the interpreter does not like it when I pass it like this. Maybe this is the reason why there is a broken paper error in that other example.

Nope. Nevermind.

10am. Let me finally run that thing. I am done studying it for the time being.

10:10am. Ok, I ran it. I am really hoping for something more interactive though.

What I really need to learn now is not the deep learning related stuff, but how to manage the environments and do data munging.

The whole plt freezing thing is really annoying. I'd like it for starters if I could find a way to resolve that. I'd like it if I could move the cartpole on my own using the keyboard.

I want to train a net and see how it does in real time.

10:20am. I am thinking of what I should do now.

https://gym.openai.com/docs/

I should dive into the Gym next and solve some of the environments using tabular methods. I need to do this as an exercise in order to get familiar with Python and I'd rather put off messing with deep nets for a bit.

This will also allow me to tackle various auxiliary challenges such as resolving the plt freezing issue. I'd be better off just replacing it with something that works properly.

10:25am. Let me get my focus on. I'll go through the docs for Gym and play with it after that."