"1:50pm. Let me start.
I have to start otherwise the sheer existential dread I feel right now will overwhelm me.

Right now I am thinking of what I should do. First thing first, I really want to solve the issue with the plot window freezing. Then I'd like to hook the environments to the keyboard to get a sense for it.

2pm. Hmmm...

2:05pm. I am a bit undecided right now.

Let me put stable-baselines on hold for the time being until I get the environments themselves in order.

2:20pm. https://github.com/openai/gym/issues/741
https://github.com/openai/gym/issues/762

This is so annoying. There are like dozens of people each groping in the dark here. What is this bullshit?

Let me take a different tack. I will just step through it all with a debugger. That will give me a feel for everything that is going on in the program I am running. Since Python is dynamic and interactive this sort of thing should not be a problem.

2:30pm. Right now I am looking through the VS debugger information.

https://www.quora.com/What-is-the-best-programming-language-to-use-when-writing-a-compiler-e-g-ML-Lisp-Java-C++-Python-et-cetera

...And this.

2:35pm. Focus, that is useless. Right now I am looking through the `PersistentVector` implementation and thinking ...nah. I do not have the tools to create a quality implementation very easily.

Rather than `PersistentVector`s, I'd be much better off with a linked list of immutable arrays for the sake of closures. And I'd do well to lay off all the partial application. I should make it so that closures track on their own how many times they've been partially applied.

2:40pm. Without a doubt, I need to use plain arrays for temporaries and list of arrays for closure representations. PersitentVector would be too slow.

One thing I do not get is how flattened closure representations could work. Well probably not in an interpreter, but for compiled languages like ML they could be arranged.

2:45pm. With some optimizations, an efficient interpreter could be done even in .NET.

I'd do it like `{closure: TypedData []; start: int; end: int} list`. No actually I could be better than this by precalculating the outer and inner scopes. It would be trivial.

A flattened representation would just be an array.

2:55pm. In that case I'd have to do more work during the prepass - every variable would have a local (live) and a global id, and I'd have to make sure to do renaming properly. I think the nested closure would be more efficient, but both would do somehow. The first one would be really simple to implement though and would be space efficient. I have no reason not to go for it...

...Forget it for now.

3:05pm. The debugger. I need to figure out how to step through the entirety of the program.

3:30pm. https://stackoverflow.com/questions/54160505/how-to-make-the-debugger-step-through-the-whole-program

```
    import gym
    env = gym.make('CartPole-v0')
    env.reset()
    for _ in range(100):
        env.render()
        env.step(env.action_space.sample()) # take a random action

This is an example from the [OpenAI gym](https://gym.openai.com/docs/). 'Step into' works neither on the `import gym` nor on any of the method calls for `env`, it just skips over them to the next statement. I am new to Python, but not to programming at this point.

In order to build intuition on how exactly `gym` and various Python libraries work I'd like as an exercise to go through them step by step. There is also a plotting issue in all the PyTorch examples and now this one where the window freezes after being run that has something to do with async handling.

Since I'd like to use Python specifically for its interactivity, I want to look into this. How could this be done?
```

I'll see what I'll get.

3:35pm. Damn, this whole Python thing is off to a rocky start. But there is no way around it. If I want to be effective at this I am going to have to master the tools.

Python is not Haskell - I am going to have to get used to stepping through the program wherever it is applicable and get used to reading other people's code. This is a significant change from how I've been doing programming up to now."