"10:25am. https://docs.python.org/3/library/tkinter.html
I am still thinking what I want exactly.

Yesterday I had the idea of redoing the cartpole problem. That is in fact what I should do.

It seems simple, but doing it properly will require me to master a wide variety of things like simple GUIs, async programming and get familiar with OpenAI's codebase, and in addition to that PyTorch. On the surface the goal would be to solve Cartpole using tabular methods, but the purpose of the journey would be to get a feel for everything that I will need.

If I am going to make Spiral an embedded language for Python at some point, this is something I'd need to do anyway.

Getting to a solid level of familiarity with all the libraries that I will need to do RL effectively in Python will be a month or two of solid effort so I won't get discouraged too easily at the rocky first impressions. It is not because of Python, but because Windows is a second class citizen when it comes to ML.

...Though I am not exactly impressed by the quality of the codebase either.

That having said, I am in a open minded mood here. Maybe it is high quality and I am just letting my prejudices cloud my mind.

Whether cartpole is high quality or not is something I'll find out when I give it a try at refactoring it.

10:40am. For now, let me play with tkinter for a bit."