"In Spiral I'd do this using CPS though.
3:25pm.

```
The first statement of the function body can optionally be a string literal; this string literal is the function’s documentation string, or docstring. (More about docstrings can be found in the section Documentation Strings.) There are tools which use docstrings to automatically produce online or printed documentation, or to let the user interactively browse through code; it’s good practice to include docstrings in code that you write, so make a habit of it.
```

The pervasive use of partial application in Spiral would make this a bit more difficult, but I think once I get all the implementation bits out of the way, figuring out how to make it more ergonomic should be a priority.

```
The execution of a function introduces a new symbol table used for the local variables of the function. More precisely, all variable assignments in a function store the value in the local symbol table; whereas variable references first look in the local symbol table, then in the local symbol tables of enclosing functions, then in the global symbol table, and finally in the table of built-in names. Thus, global variables cannot be directly assigned a value within a function (unless named in a global statement), although they may be referenced.
```

Ah, I see. That explains the function in the GUI example. It can access the global variables, but not modify them without the `global` declaration.

3:40pm.

```
Important warning: The default value is evaluated only once. This makes a difference when the default is a mutable object such as a list, dictionary, or instances of most classes. For example, the following function accumulates the arguments passed to it on subsequent calls:

def f(a, L=[]):
    L.append(a)
    return L

print(f(1))
print(f(2))
print(f(3))

...

[1]
[1, 2]
[1, 2, 3]
```

Need to keep this in mind.

3:55pm. Damn, this is boring. It sure is boring.

```
Remove the item at the given position in the list, and return it. If no index is specified, a.pop() removes and returns the last item in the list. (The square brackets around the i in the method signature denote that the parameter is optional, not that you should type square brackets at that position. You will see this notation frequently in the Python Library Reference.)
```

Need to keep the optional argument notation in mind.

4:10pm. `help` is rather helpful. It will actually print out all the methods and data descriptors of an object. This was not covered by the tutorial yet, I remembered it from my previous experience with Python.

4:25pm. "Note that in Python, unlike C, assignment cannot occur inside expressions. C programmers may grumble about this, but it avoids a common class of problems encountered in C programs: typing = in an expression when == was intended."

This is good design, though I'd prefer it if the whole language was expression based like F# and Spiral. Otherwise exceptions like this will lead to complexity in the language specification.

4:35pm. "This imports all names except those beginning with an underscore (_). In most cases Python programmers do not use this facility since it introduces an unknown set of names into the interpreter, possibly hiding some things you have already defined."

I assumed that `from fibo import *` imports everything. Good thing I am going through the documentation. But damn this is boring.

4:55pm. https://docs.python.org/3/tutorial/modules.html

Wow, damn this is boring. To think I am only at a third of it now.

Yeah, it is easy, but to make up for it the boredom comes in.

"Python checks the modification date of the source against the compiled version to see if it’s out of date and needs to be recompiled."

Ah, I am such a idiot. I was wondering how I could check whether a file needs recompilation when implementing caching in Spiral. To think it would be this. It never even occurred to me to check the date.

5pm. Ah, `dir`. Maybe I could use this to check all the variables currently in scope.

```
def f():
    a = 2
    def g():
        b = 3
        print(dir())
    g()

dir()
f()
```

No this just print the variables in local scope.

Well, there is bound to turn up something.

5:05pm.

```
The __init__.py files are required to make Python treat the directories as containing packages; this is done to prevent directories with a common name, such as string, from unintentionally hiding valid modules that occur later on the module search path. In the simplest case, __init__.py can just be an empty file, but it can also execute initialization code for the package or set the __all__ variable, described later.
```

All this is required knowledge, but I sure am yawning here.

5:10pm. https://docs.python.org/3/tutorial/inputoutput.html

Just this thing and then I think I will stop for the day. The boredom of this is getting to me.

5:30pm. https://docs.python.org/3/tutorial/errors.html

Let me stop here. The sheer tedium of this is killing me.

It is time to play.

I am not sure whether I will have time to test the RL scheme that in the next month.

Learning Python, redoing the Cartpole and maybe working on Spiral will be quite enough to fill my plate.

As expected, basic programming is quite enough to fill the plate. The sheer amount of errors in PyTorch, Gym and the spelling errors stable baselines is telling me that not many people are seriously looking into ML. I think the number of posers is way out of proportion with the real players.

5:35pm. Master Spiral, master a dynamic language like Python and then take the world. That sounds like a good plan.

I think I'll just study Python for a few days and then work on Spiral for a few days and go back and forth between the two. It depends on whether I've settled on the design. If I could go for a few days without changing my mind then that will be a signal to go ahead."