"4:25pm. https://docs.python.org/3/reference/compound_stmts.html#async-def
Ah, Python has plenty of async functionality. This was not covered in the tutorial.

4:30pm.

```
and TypedExpr = // TypedData being `TyList []` indicates a statement.
    | TyOp of TypedData * ConsedTy * Trace * Op * TypedData
    | TyIf of TypedData * ConsedTy * Trace * tr: TypedExpr [] * fl: TypedExpr []
    | TyCase of TypedData * ConsedTy * Trace * TyTag * (TypedData * TypedExpr []) []
    | TyJoinPoint of TypedData * ConsedTy * Trace * JoinPointKey * JoinPointType * CallArguments
    | TyLocalReturn of TypedData * ConsedTy * Trace
```

The language reference is so boring that my mind went back to this.

Actually, let me go for this. The return should have a special role after all rather than being an op. It will make things easier in the codegen as now `TypedData` will always indicate the proper return. Return should have a type and not be a statement. Statements are presumed to be of unit type.

4:45pm. Ok, the language spec is a bit too dry for me. Or maybe I am just tired at this point.

4:55pm. https://docs.python.org/3/reference/executionmodel.html#interaction-with-dynamic-features

I might need to look into this. This stuff on name resolution and how Python has eval might turn up something for me.

https://www.reddit.com/r/learnprogramming/comments/w6i0p/bytecode_execution_vs_ast_walking_speed/?st=jqwjgloq&sh=6b6a959b

"Here are some rather dated metrics along those lines:

http://picolisp.com/5000/!wiki?NeedForSpeed
Picolisp is one of the few actual interpreters that I'm aware of.

The former ones are slower, because AST walking involves recursion and virtual calls while bytecode can be executed in a loop.

Nah. :)

You're making a lot of assumptions there, about how an interpreter has to be implemented, and also about the cost of implementation techniques in the host language.

On the implementation side, you can build a tree-based interpreter as single loop containing a case statement, and using a pair of arrays to act as code and data stacks.

On the host language side, proper tail recursion needn't cost more than a single JMP instruction at the machine code level, so it can be exactly as efficient as a loop.

Those are just two examples. It's kind of a big subject, but basically, there a are a lot of ways to make a tree-based interpreter as fast or faster than a byte-code interpreter for the same language.

On the other hand, if you said that building an AST from source text and then walking it is slower than running pre-compiled-and pre-optimized bytecode, then that would be a lot harder to argue with."

I am getting distracted like a donkey.

Let me finish the module chapter of the language reference and then I'll stop for the day.

5:30pm. Thunder. I need to turn off.

6:50pm. Done with lunch. Let me see if the manual is still in the cache. I won't turn on the Internet just yet.

Let me watch some anime. Kotobuki and Kemurikusa are on the backlog.

7:20pm. I quite like Kotobuki. Let me make that commit now.

I had to run before I could finish the day properly.

Tomorrow I will hopefully go through all the Python documentation.

Chapter 4 of the language spec was especially interesting to me - if I want to add Spiral to Python, I'll have to see exactly what it can do using `exec` and `eval` functions. When I tried dir and saw that it does not capture everything from the scope I was discouraged, but there might be a way to get what I want without necessarily having to build a parser on top.

7:30pm. I am going to die if I continue writing this so let me just stop here for the day."