"9:15am. Whenever I miss a day, it means that a tragedy has struck. Yesterday, my monitor died.
I spent the day squatting at my dad's laptop and reading Python tutorials, waiting for that VGA cable to arrive so I could make use of an old monitor I had lying around. The GPU needed to be disconnected from the motherboard to make this happen. Well, it is not like I am doing any heavy duty ML work so it fine. The VGA -> HDMI adapter should arrive at some point.

9:25am. It is not important. Yesterday was a real brush with mortality.

Suppose that computer itself instead of the monitor had died, what would I have done?

Started applying for jobs. At that point I'd started praying that the work on Spiral made me at least somewhat employable. I'd make a post on the next PL thread along with the update asking whether anybody wanted to hire me and started getting creative with my resume.

Right now I feel the pressure to do something with the poker plan.

Spiral is feature complete in terms of the front end, though I need to redo the backend. I've given up on thinking that a systems language like it will be enough to carry me all the way. It is true that intensional is more expressive than extensional, but the price for that is more severe than I ever imagined.

More than that while it is true that is ridiculous to expect to cause the Singularity when you cannot even make a generic Cuda map kernel, it is also ridiculous to expect to be able to cause it from the command line. The compile-edit loop that I am using is ridiculous. It might be good enough to get a language off the ground, but that is in no way the way to program. Spiral has a very important role as the backbone, but it cannot be the muscle.

My sense is that Spiral (once I've speed it up) will have a very important role, but I need more.

I thought that a dynamic language like Python might be what I am looking for, but I am not sure it has what it takes for reasons I will go into shortly."