"9:40am. https://stackoverflow.com/questions/459083/how-do-you-run-your-own-code-alongside-tkinters-event-loop
As a exercise I am considering Python from the interactivity standpoint.

9:50am. As a rough measure, how good a language is at GUI programming is how good it is in general. Just why is it so difficult to spin the GUI off on its own thread and communicate using message passing? This should be the standard approach. Why is that PyTorch code sprinkled with sleep statements and why do the plotting windows freeze?

I think it should be common sense that if you are using a interactive dynamic language as opposed to a systems language like Spiral, that GUIs should be first class.

Yeah, it is true that either way Python has better support for interactive programming, but it could be much better. I should not skulk through old SO posts just to figure out how to make concurrency happen for a GUI.

Quite frankly I am starting to sense that I cannot make all that I want happen. Even if I went with Python, it would not future proof me.

What I want is something that would allow me to compose interpreters.

For a real AI language, one thing I would want is that the language itself would have a notion of memory. Like, I would not expect to write a script to save and load models on my own, rather they should be a part of the image. The interpreter itself should take of all the serialization and deserialization. I should be able to edit it as I go along.

The model needs to be a part of the program itself. This needs to be recognized as it certainly greatly influences the evolution of it, even it is not a part of it directly like in Python and Spiral.

At some point AIs will be doing lifelong learning. Can you really imagine managing that process using Python scripts? Of course not.

10:05am. I need more, so much more than this.

When I turn off for the day, the agents should turn off as well. All of this needs to be automated.

For what I want, really only the Smalltalk languages from what I've seen fit the bill. So maybe at this critical juncture, I should consider Pharo as my workhorse.

That having said, before I decide I want to study a bit more. Maybe Pharo will turn out to be shitty. Maybe Ruby, it's popular relative will fit the bill.

I also want to learn a bit more about PicoLisp. I find that I am more interested in highly opinionated languages as a programmer than those who try to straddle the middle.

Also, I starting to think that I should shift my goals. I mean even if I could put in a great effort just to hack out some poker agent in Python + PyTorch does not mean I should. I should do it properly. And if I am desperate for money, I should just get a job.

It would certainly be a better plan than trying to do the immense undertaking of assaulting the online gambling dens before the pieces for it are in place. I do not want to be a bottom feeder that just plays a large number of hands - I seek absolute victory and the destruction of my enemies. Anything less is cowardice. And I sure have been a coward for the past two weeks.

Right now it is just late 2016. The shadow of it has been resurrected and I once more have a juncture in front of me.

Do I compromise here and go with Python, or do I aim for the final time to grasp everything and achieve absolute mastery of programming?

If I was the kind of person to compromise I would have never set out on this path in the first place. For me the course should be clear. I aim for the highest peak and risk absolute destruction.

10:25am. I will do this properly. I will absolutely do this properly. ML needs to be done in a language suitable for it rather than the thing convenient for it.

10:30am. I'll give up on making money with it. Should the need for it arise, I will just make up my mind and look for paid work.

10:35am. I am an idiot. I am totally a researcher at this point, but have the self image of a practical guy. Let's say I am a practical researcher and leave it at that.

Instead of wasting time ranting, I think I'll study PicoLisp and Ruby. Since Ruby is kind of large, I'll just look at how GUIs in it operate."