"https://www.reddit.com/r/Python/comments/7rp4xj/threading_a_tkinter_gui_is_hell_my_least_favorite/?st=jqyzruzi&sh=f461dc77
11:10am. http://shoesrb.com/

This looks really polished. As expected OO is ideal for message passing concurrence. There is actually very little difference between calling methods and passing messages. A pure OO language should be ideally positioned in order to take advantage of this.

https://saveriomiroddi.github.io/An-overview-of-ruby-gui-development-in-2018/

Or maybe not. This says that the functionality of Shoes is really poor.

11:20am. http://books.pharo.org/spec-tutorial/

Let me take a break here. I'll read this later.

12:10pm. Mini Toji time. Then I'll go through the PicoLisp documentation.

12:35pm. https://picolisp.com/wiki/?metaprogramming

To be honest, PicoLisp is pretty ugly as a language so I won't be using it, but I am interested in how it handles fexprs.

12:40pm. https://picolisp.com/wiki/?firstclassenvironments

So it does use dynamic binding even now.

12:45pm. Forget it. I am not interested in an ugly language that tries to pass off leakiness as expressiveness.

http://files.pharo.org/books-pdfs/spec/2017-01-23-SpecBooklet.pdf

Before I move on to reading this, let me actually watch the talk on Spec. I am not interested in reading that much right now."