"3:20pm. Focus me. Stop reading the ML sub.
https://www.infoworld.com/article/2870966/ruby/ruby-pioneers-come-clean-on-languages-shortcomings.html

4pm. 37/84. I did not publish it on Github, but I did a poker app back in 2016. It was a huge pain to do the GUI imperatively, so it is really nice to see it done functionally here. I like how everything is an expression in Smalltalk.

4:20pm. http://mooc.pharo.org/
http://books.pharo.org/

I think I should go through the MOOC and then the `Pharo By Example` book. I sort of get Pharo on a higher level, but not internally. With the Python it is easy as it is so similar to languages I already know, but with Pharo, I need some experience in order to make it natural.

I have a choice between Python and Pharo, but I really like the later much more. I liked it the most of all the dynamic languages that I've seen so far.

4:40pm.

```
Character cr.

Character space.
```

I am having difficulty figuring out what `cr` and `space` are here. For some reason they do not show up in the inspector.

4:50pm. I somehow did an access to an invalid memory location. I am not sure how I triggered it. I was messing with the color class.

4:55pm. It seems that pressing Ctrl + P/O/I (each key in succession) causes `LoadLibrary(SurfacePlugin.dll) (998: Invalid access to memory location.` to trigger.

5:45pm. I've gone through the ProfStef tutorial.

5:55pm. Damn, I wish I had not erased all the videos when I was clearning the downloads folder. I had all the MOOC files on drive.

Well, I'll just get it again.

https://www.youtube.com/channel/UCp3mNigANqkesFzdm058bvw
Pharo Project talks

I kind of feel like getting into it.

http://books.pharo.org/updated-pharo-by-example/pdf/2018-09-29-UpdatedPharoByExample.pdf
Pharo By Example

Rather than go through the MOOC again, how about I just go through the book instead.

6:40pm. https://www.youtube.com/watch?v=xhPlUaXpCU4
Pharo: Objects at your Fingertips

This talk is old now. But let me go for it.

https://youtu.be/xhPlUaXpCU4?t=1355

Actually this is quite interesting. They can do assembly in Pharo. I am guessing they are just using the OO syntax in order to build it up.

7:10pm. https://youtu.be/xhPlUaXpCU4?t=1688

Ah, these instance variables are there purely for the sake of IDE it seems.

https://youtu.be/xhPlUaXpCU4?t=1731

"It is really interesting. At the beginning you think 'Oh yeah, boy that can be handy when I debug' or something like that. But over time you start to really change the way you develop completely so you stop thinking about what might happen, but you always look at what happens. So you always change something, try out, look into the object, see that it works or not and change it again, and up to the point where you go back to another system you start to feel really, really constrained. Because there are no compile time here. It is not that when we edit we need to compile, restart, not at all. We can even keep our objects live and they will change with the code a little bit to some extent. Add instance variables and get the instance variables in the object and things like that.

So it is very nice and after a while it is really a completely different way of really working."

Yeah, I bet. This is why I am interested in Pharo in the first place. I could go further, much further in it than I could with Python.

7:35pm. https://youtu.be/xhPlUaXpCU4?t=2699

Images go into terrabyte size. This is something I'd want for NNs. I've been really wondering about this.

Now that the talk is over it is interesting how much he emphasized the importance of the IDE.

https://www.youtube.com/watch?v=hBd257eRIq4&list=PL4actYd6bfnx8l8cqYWhdNB6gJgRvk5sH
Artificial Intelligence (Neural Networks)

There is specifically a talk on Pharo about this. Let me go for it for a bit and then I will call it a day.

7:45pm. Let me call it a day. I am not interested in a super basic talk on NNs."